pyproject.toml
[tool.poetry]
name = "wainstitution-yrb-2021"
version = "0.1.0"
description = "Water quota influences in the yellow river basin"
authors = ["Shuang Song <songshgeo@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "src", from='core'}, {include = "core"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
# NOT compitible package `qdldl`: https://github.com/osqp/qdldl-python/issues/25
# syntheticcontrolmethods = "^1.1.17"
cos = { git = "git@github.com:SongshGeo/tencent-cos-manipulations.git", branch = "dev" }
# For help: https://www.cvxpy.org/install/index.html#install-from-source
cvxpy = {git = "git@github.com:cvxpy/cvxpy.git", branch = "master"}  # Install from source for MacOS M1 sillcon version
pca = "^1.9.1"
pandas = "^1.5.3"
rasterio = "^1.3.6"
geopandas = "^0.12.2"
tqdm = "^4.64.1"
prettytable = "^3.6.0"
numpy = "^1.24.2"
matplotlib = "^3.7.0"
pygam = "^0.8.0"
xarray = "^2023.2.0"
pyyaml = "^6.0"
bidict = "^0.22.1"
seaborn = "^0.12.2"
pingouin = "^0.5.3"
mksci-font = "*"
scipy = "^1.4.1"
ecos = "^2.0.12"
syntheticcontrolmethods = "^1.1.17"
openpyxl = "^3.1.1"
hydra-core = "^1.3.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
allure-pytest = "^2.12.0"
flake8 = "^6.0.0"
isort = "^5.12.0"
nbstripout = "^0.6.1"
pydocstyle = "^6.3.0"
pre-commit-hooks = "^4.4.0"
pre-commit = "^3.1.0"
ipykernel = "^6.21.2"
notebook = "^6.5.2"
jupyterlab = "^3.6.1"
jupyterlab-execute-time = "^2.3.1"
jupyterlab-lsp = "^3.10.2"
python-lsp-server = "^1.7.1"
jupyter = "^1.0.0"
black = "^23.1.0"
hydra-core = "^1.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
